refactor: streamline catalog linking and synchronization process by consolidating error handling and moving CatalogProductInsertion to a dedicated module
